refactor: Omit 사용하는 대신 FormValue 타입 생성 후 사용
